+++ pinnwand/error.py
@@ -10,3 +10,10 @@ class RatelimitError(ValueError):
     ratelimit for a certain action."""
 
     pass
+
+
+class SpamError(ValueError):
+    """This exception is used to indicate that a user has exceeded the allowed
+    spamscore."""
+
+    pass